feat(util): better file utils
